# app/docs/api.yml
openapi: 3.0.3
info:
  title: Mini Attack Surface Management API
  description: |
    API cho hệ thống quản lý Attack Surface Management.
    Phát triển qua 6 buổi học (3 tuần, mỗi tuần 2 buổi).
  version: 1.0.0
  contact:
    name: Intern Program

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /health:
    get:
      summary: Health check
      description: Kiểm tra trạng thái service và database connection
      operationId: healthCheck
      tags:
        - System
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  database:
                    type: string
                    example: connected
                    description: Database connection status (từ buổi 3)

  /assets:
    post:
      summary: Create a new asset
      description: Tạo mới một asset (domain, IP, hoặc service)
      operationId: createAsset
      tags:
        - Assets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: example.com
                  description: Tên của asset
                type:
                  type: string
                  enum: [domain, ip, service]
                  example: domain
                  description: Loại asset
                status:
                  type: string
                  enum: [active, inactive]
                  default: active
                  example: active
                  description: Trạng thái asset (optional, default là active)
      responses:
        "201":
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    get:
      summary: List all assets
      description: |
        Lấy danh sách tất cả assets với khả năng filter và search.
        - Buổi 2: Basic list all
        - Buổi 4: Thêm filtering và searching
      operationId: listAssets
      tags:
        - Assets
      parameters:
        - name: type
          in: query
          description: Filter theo loại asset
          required: false
          schema:
            type: string
            enum: [domain, ip, service]
        - name: status
          in: query
          description: Filter theo trạng thái
          required: false
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          description: Tìm kiếm theo tên (partial match)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /assets/{id}:
    get:
      summary: Get a single asset
      description: Lấy thông tin chi tiết của một asset theo ID
      operationId: getAsset
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của asset
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Asset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      summary: Update an asset
      description: Cập nhật thông tin của một asset
      operationId: updateAsset
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của asset
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  description: Tên mới của asset
                type:
                  type: string
                  enum: [domain, ip, service]
                  description: Loại asset mới
                status:
                  type: string
                  enum: [active, inactive]
                  description: Trạng thái mới
      responses:
        "200":
          description: Asset updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete an asset
      description: Xóa một asset khỏi hệ thống
      operationId: deleteAsset
      tags:
        - Assets
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của asset
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Asset deleted successfully
        "404":
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Asset:
      type: object
      required:
        - id
        - name
        - type
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
          description: Unique identifier (UUID)
        name:
          type: string
          example: "example.com"
          description: Tên của asset
        type:
          type: string
          enum: [domain, ip, service]
          example: "domain"
          description: Loại asset
        status:
          type: string
          enum: [active, inactive]
          example: "active"
          description: Trạng thái của asset
        created_at:
          type: string
          format: date-time
          example: "2026-03-02T10:30:00Z"
          description: Thời điểm tạo
        updated_at:
          type: string
          format: date-time
          example: "2026-03-02T15:45:00Z"
          description: Thời điểm cập nhật gần nhất

    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid input"
          description: Mô tả lỗi